backend/requirements.txt
fastapi
mysql-connector-python
bcrypt
pyjwt